backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
pillow==12.0.0
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1